+++ index.js
@@ -34,7 +34,8 @@ import {read} from "riff-chunks";
 import {decode as decodeADPCM, encode as encodeADPCM} from "imaadpcm";
 import {alaw, mulaw} from "alawmulaw";
 import {encode, decode} from "base64-arraybuffer";
-import {pack, unpack, packArray, unpackArray, types} from "byte-data";
+import {pack, unpack, packArray, unpackArray,
+    unpackFrom, unpackArrayFrom, types} from "byte-data";
 
 /**
  * @type {!Object}
@@ -1569,7 +1570,7 @@ class WaveFile {
                     item["dwDialect"] = this.read_(subChunk["chunkData"], uInt16_);
                     item["dwCodePage"] = this.read_(subChunk["chunkData"], uInt16_);
                 }
-                item["value"] = this.readZSTR_(subChunk["chunkData"].slice(this.head_));
+                item["value"] = this.readZSTR_(subChunk["chunkData"], this.head_);
                 this.LIST[this.LIST.length - 1]["subChunks"].push(item);
             }
         // RIFF 'INFO' tags like ICRD, ISFT, ICMT
@@ -1577,8 +1578,9 @@ class WaveFile {
             this.LIST[this.LIST.length - 1]["subChunks"].push({
                 "chunkId": subChunk["chunkId"],
                 "chunkSize": subChunk["chunkSize"],
-                "value": this.readZSTR_(subChunk["chunkData"].slice(0))
+                "value": this.readZSTR_(subChunk["chunkData"],  0)
             });
+
         }
     }
 
@@ -1605,14 +1607,14 @@ class WaveFile {
      * @return {string} The string.
      * @private
      */
-    readZSTR_(bytes) {
+    readZSTR_(bytes, index) {
         /** @type {string} */
         let str = "";
-        for (let i=0; i<bytes.length; i++) {
+        for (let i=index; i<bytes.length; i++) {
             if (bytes[i] === 0) {
                 break;
             }
-            str += unpack([bytes[i]], types.chr);
+            str += unpackFrom(bytes, types.chr, i);
         }
         return str;
     }
@@ -1628,7 +1630,7 @@ class WaveFile {
         /** @type {string} */
         let str = "";
         for (let i=0; i<maxSize; i++) {
-            str += unpack([bytes[this.head_]], types.chr);
+            str += unpackFrom(bytes, types.chr, this.head_);
             this.head_++;
         }
         return str;
@@ -1645,8 +1647,7 @@ class WaveFile {
         /** @type {number} */
         let size = bdType["bits"] / 8;
         /** @type {number} */
-        let value = unpack(
-            bytes.slice(this.head_, this.head_ + size), bdType);
+        let value = unpackFrom(bytes, bdType, this.head_);
         this.head_ += size;
         return value;
     }
@@ -1706,6 +1707,8 @@ class WaveFile {
     samplesFromBytes_(bytes) {
         this.data.samples = unpackArray(
             bytes, this.getSamplesType_());
+        //this.data.samples = unpackArrayFrom(
+        //    bytes, this.getSamplesType_());
     }
 
     /**